HiredAI/requirements.txt
langchain
langchain-community
pypdf
langchain-ollama
selenium
streamlit
beautifulsoup4
pydantic
streamlit
openai
requests
pandas
numpy
scikit-learn
python-dotenv
PyPDF2
tqdm
speechrecognition
gtts
pydub

